+++ profiles/stac-server/inputs.tf
@@ -60,6 +60,27 @@ variable "stac_server_inputs" {
     cors_headers                                = string
     authorized_s3_arns                          = list(string)
     api_rest_type                               = string
+    api_lambda = optional(object({
+      handler         = optional(string)
+      memory_mb       = optional(number)
+      runtime         = optional(string)
+      timeout_seconds = optional(number)
+      zip_filepath    = optional(string)
+    }))
+    ingest_lambda = optional(object({
+      handler         = optional(string)
+      memory_mb       = optional(number)
+      runtime         = optional(string)
+      timeout_seconds = optional(number)
+      zip_filepath    = optional(string)
+    }))
+    pre_hook_lambda = optional(object({
+      handler         = optional(string)
+      memory_mb       = optional(number)
+      runtime         = optional(string)
+      timeout_seconds = optional(number)
+      zip_filepath    = optional(string)
+    }))
     auth_function = object({
       cf_function_name             = string
       cf_function_runtime          = string
@@ -106,6 +127,9 @@ variable "stac_server_inputs" {
     cors_headers                                = ""
     authorized_s3_arns                          = []
     api_rest_type                               = "EDGE"
+    api_lambda                                  = null
+    ingest_lambda                               = null
+    pre_hook_lambda                             = null
     auth_function = {
       cf_function_name             = ""
       cf_function_runtime          = "cloudfront-js-2.0"